backend/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-Cors==4.0.0
python-dotenv==1.0.0
Authlib==1.2.0
requests==2.31.0
